Added, generating straight line
Generating straight line was achived by using the Bresenham's line algorithm which allows to draw line between two points in a discrete grid. This algorithm is used in generating the pictures. This implementation allows to use starting positions, end positions and number of steps. This approach will be ideal to generating straight line to comparing to previously generate trajectories