Added new version of function generating points in line
Apperently the current approach dont provide the possibility to generate the points in line constructed of 2 points with the same x values due to the problem when slope was generating division by zero error. New version provides the check if the x coordinates are the same and generate y points accordingly also ensuring that the starting and end positions are the same as proviaded to generate the line.